backend/pyproject.toml
[project]
name = "findemo-backend"
version = "1.0.0"
description = "Commodity Hedging & VaR Demo - Backend API"
requires-python = ">=3.11"
authors = [
    {name = "Findemo Team"}
]

dependencies = [
    # Web Framework
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    
    # Database
    "sqlalchemy[asyncio]==2.0.25",
    "asyncpg==0.29.0",
    
    # Data Validation
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    
    # Authentication
    "python-jose[cryptography]==3.3.0",
    "bcrypt==5.0.0",
    
    # File Upload
    "python-multipart==0.0.6",
    
    # Excel Processing
    "openpyxl==3.1.2",
    "pandas==2.2.0",
    
    # Math & Science
    "numpy==1.26.3",
    "scipy==1.12.0",
    
    # Market Data
    "yfinance==0.2.35",
    "pandas-datareader==0.10.0",
    
    # HTTP Client
    "httpx==0.26.0",
]

[project.optional-dependencies]
test = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.3",
    "psycopg2-binary==2.9.9",
]

dev = [
    "black==24.1.1",
    "ruff==0.1.14",
    "mypy==1.8.0",
    "ipython==8.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
